Factor out common functions for upcoming llloop.py update
